+++ packages/editor/editor-core/src/plugins/block-type/ui/ToolbarBlockType/styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { HTMLAttributes, ComponentClass } from 'react';
+
+export const BlockTypeMenuItem: ComponentClass<
+  HTMLAttributes<{}> & {
+    tagName: string;
+    selected?: boolean;
+  }
+> = styled.div`
+  ${props => (props.selected ? `${props.tagName} { color: white }` : '')};
+`;