+++ packages/editor/editor-mobile-bridge/src/renderer/mobile-renderer-element.tsx
@@ -12,7 +12,7 @@ export interface MobileRendererState {
   document: any;
 }
 
-(window as any).rendererBridge = new RendererBridgeImpl();
+const toNativeBridge = ((window as any).rendererBridge = new RendererBridgeImpl());
 
 export default class MobileRenderer extends React.Component<
   {},
@@ -40,6 +40,14 @@ export default class MobileRenderer extends React.Component<
     });
   }
 
+  private onLinkClick(url) {
+    if (!url) {
+      return;
+    }
+
+    toNativeBridge.onLinkClick(url);
+  }
+
   render() {
     try {
       // If we haven't received a document yet, don't pass null.
@@ -54,6 +62,17 @@ export default class MobileRenderer extends React.Component<
           dataProviders={this.providerFactory}
           appearance="mobile"
           document={this.state.document}
+          eventHandlers={{
+            link: {
+              onClick: (event, url) => {
+                event.preventDefault();
+                this.onLinkClick(url);
+              },
+            },
+            smartCard: {
+              onClick: this.onLinkClick,
+            },
+          }}
         />
       );
     } catch (ex) {